Update to :gem: Unbound v1.20.0 and :lock: OpenSSL v3.0.13
Updated Unbound to v1.20.0. Updated OpenSSL to v3.0.13 for Unbound
build. Updated Docker UBI to Debian Bookworm. Updated docker-compose
to mount custom stubby.yml and a-records.conf. Renamed a-records.conf
to a-records.conf-example and updated readme to reflect the change.
For custom A records and CNAMEs, rename the example to .conf and
update as needed.